+++ R/functions_spatial.R
@@ -12,7 +12,8 @@
 #'   then a random sample of \code{x} will be taken of the specified size.
 #' @inheritParams as_points
 #' @param seed Passed to \code{\link{set.seed}}; used only if
-#'   \code{sub_samplepoints_N} is a number.
+#'   \code{sub_samplepoints_N} is a number and \code{seed} is not \code{NULL}.
+#'   The random number state of the caller is restored on exit.
 #'
 #' @seealso \code{\link[automap]{autofitVariogram}}
 #'
@@ -51,7 +52,20 @@ variogram_range <- function(
   pts[, "target"] <- 1
 
   if (!is.null(sub_samplepoints_N)) {
-    set.seed(seed)
+    if (!is.null(seed)) {
+      # Restore the random number state of the caller on exit
+      oldSeed <- get0(".Random.seed", envir = globalenv(), inherits = FALSE)
+      on.exit(
+        if (is.null(oldSeed)) {
+          rm(".Random.seed", envir = globalenv())
+        } else {
+          assign(".Random.seed", oldSeed, envir = globalenv())
+        },
+        add = TRUE
+      )
+      set.seed(seed)
+    }
+
     tmp <- sample.int(
       n = nrow(pts),
       size = sub_samplepoints_N,